// frontend/src/api/client/models/Body_upload_ligand_to_target_api_v1_drug_discovery_upload_ligand_to_target_post.ts
/* generated using openapi-typescript-codegen -- do no edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type Body_upload_ligand_to_target_api_v1_drug_discovery_upload_ligand_to_target_post = {
    experiment_id: string;
    target_id: string;
    sdf_file: Blob;
};